fastapi/requirements.txt
annotated-types
anyio
cachetools
certifi
charset-normalizer
click
colorama
fastapi
faiss-cpu
sentence-transformers
google-auth
redis
orjson
google-genai
h11
httpcore
httpx
idna
pip
pyasn1
pyasn1_modules
pydantic
pydantic_core
requests
rsa
sniffio
starlette
tenacity
typing_extensions
typing-inspection
urllib3
uvicorn
websockets
